+++ packages/contracts/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- Make `performOperation` nonReentrant
 - Remove unnecessary `_nonce` state variable, use current tree size as nonce
 - Remove `opSuccess` and add `opProcessed`, which indicate internal error during processing
 - Fix asset/id encoding according to design doc